backend/pyproject.toml
[project]
name = "rag-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi",
    "uvicorn",
    "python-multipart",
    "httpx",
    "python-dotenv",
    "pydantic",
    "asyncpg>=0.30.0",
    "openai>=1.78.0",
]
